+++ src/errors.ts
@@ -1,0 +1,6 @@
+export class PlattformUnsupportedError extends Error {
+    override name: "PlattformUnsupportedError" = "PlattformUnsupportedError";
+    constructor(msg?: string) {
+        super(msg);
+    }
+}