+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Add new global setting
     - Custom paket executable path
 - Add log to VS Code Output Channel
+- Added Documentation to readme
+- include paket.exe in extension instead of downloading it
 
 ## 0.2.2
 - iterate through workspaces